PiDMD update.
