+++ pydmd/rdmd.py
@@ -10,44 +10,7 @@ Systems, 18, 2019.
 import numpy as np
 
 from .cdmd import CDMD
-
-
-def compute_rank(X, svd_rank=0):
-    """
-    Rank computation for the truncated Singular Value Decomposition.
-    :param numpy.ndarray X: the matrix to decompose.
-    :param svd_rank: the rank for the truncation; If 0, the method computes
-        the optimal rank and uses it for truncation; if positive interger,
-        the method uses the argument for the truncation; if float between 0
-        and 1, the rank is the number of the biggest singular values that
-        are needed to reach the 'energy' specified by `svd_rank`; if -1,
-        the method does not compute truncation. Default is 0.
-    :type svd_rank: int or float
-    :return: the computed rank truncation.
-    :rtype: int
-    References:
-    Gavish, Matan, and David L. Donoho, The optimal hard threshold for
-    singular values is, IEEE Transactions on Information Theory 60.8
-    (2014): 5040-5053.
-    """
-    U, s, _ = np.linalg.svd(X, full_matrices=False)
-
-    def omega(x):
-        return 0.56 * x**3 - 0.95 * x**2 + 1.82 * x + 1.43
-
-    if svd_rank == 0:
-        beta = np.divide(*sorted(X.shape))
-        tau = np.median(s) * omega(beta)
-        rank = np.sum(s > tau)
-    elif 0 < svd_rank < 1:
-        cumulative_energy = np.cumsum(s**2 / (s**2).sum())
-        rank = np.searchsorted(cumulative_energy, svd_rank) + 1
-    elif svd_rank >= 1 and isinstance(svd_rank, int):
-        rank = min(svd_rank, U.shape[1])
-    else:
-        rank = min(X.shape)
-
-    return rank
+from .utils import compute_rank
 
 
 class RDMD(CDMD):
